update error, docstrings, rename
